put mock methods to k8sutil to avoid cycling imports
