+++ pkg/cluster/cluster_test.go
@@ -30,7 +30,7 @@ var cl = New(
 			},
 		},
 	},
-	k8sutil.KubernetesClient{},
+	k8sutil.NewMockKubernetesClient(),
 	acidv1.Postgresql{},
 	logger,
 )
@@ -325,6 +325,34 @@ func TestShouldDeleteSecret(t *testing.T) {
 		if outcome, username := cl.shouldDeleteSecret(tt.secret); outcome != tt.outcome {
 			t.Errorf("%s expects the check for deletion of the username %q secret to return %t, got %t",
 				testName, username, tt.outcome, outcome)
+		}
+	}
+}
+
+func TestSetStatus(t *testing.T) {
+
+	tests := []struct {
+		status  acidv1.PostgresStatus
+		outcome bool
+	}{
+		{
+			status:  acidv1.PostgresStatus{PostgresClusterStatus: acidv1.ClusterStatusCreating},
+			outcome: cl.Status.Creating(),
+		},
+		{
+			status:  acidv1.PostgresStatus{PostgresClusterStatus: acidv1.ClusterStatusRunning},
+			outcome: cl.Status.Running(),
+		},
+		{
+			status:  acidv1.PostgresStatus{PostgresClusterStatus: acidv1.ClusterStatusInvalid},
+			outcome: !cl.Status.Success(),
+		},
+	}
+
+	for _, tt := range tests {
+		cl.SetStatus(tt.status.PostgresClusterStatus)
+		if tt.outcome {
+			t.Errorf("Wrong status: %s", cl.Status.String())
 		}
 	}
 }
